// src/02-component-patterns/interfaces/Product.interface.ts
import { Props as ProductCardProps } from '../components/ProductCard';
import { Props as ProductTitleProps } from '../components/ProductTitle';
import { Props as ProductImageProps } from '../components/ProductImage';
import { Props as ProductButtonsProps } from '../components/ProductButtons';

export interface ProductContextProps {
  product: Product;
  counter: number;
  counterHandler: (value: number) => void;
}

export interface Product {
  id: string;
  title: string;
  img?: string;
}

export interface ProductCardElements {
  (Props: ProductCardProps): JSX.Element;
  Title: (Props: ProductTitleProps) => JSX.Element;
  Image: (Props: ProductImageProps) => JSX.Element;
  Buttons: (Props: ProductButtonsProps) => JSX.Element;
}